+mais exercicios

+++ desafios/desafio07-media/desafio07.py
@@ -1,5 +1,5 @@
 nome = input('Digite o nome do aluno: ')
-nota1 = int(input('Digite a primeira nota: '))
-nota2 = int(input('Digite a segunda nota: '))
+nota1 = float(input('Digite a primeira nota: '))
+nota2 = float(input('Digite a segunda nota: '))
 media = (nota1 + nota2) / 2
 print('A media do aluno {} e {}'.format(nome, media))
